+++ code/stable/glassbr/src/cpp/Makefile
@@ -12,8 +12,8 @@ endif
 
 build: GlassBR$(TARGET_EXTENSION)
 
-GlassBR$(TARGET_EXTENSION): InputParameters.hpp DerivedValues.hpp InputConstraints.hpp InputFormat.hpp OutputFormat.hpp Calculations.hpp ReadTable.hpp Interpolation.hpp Control.cpp InputParameters.cpp DerivedValues.cpp InputConstraints.cpp InputFormat.cpp OutputFormat.cpp Calculations.cpp ReadTable.cpp Interpolation.cpp
-	"$(CXX)" Control.cpp InputParameters.cpp DerivedValues.cpp InputConstraints.cpp InputFormat.cpp OutputFormat.cpp Calculations.cpp ReadTable.cpp Interpolation.cpp --std=c++11 -o GlassBR$(TARGET_EXTENSION)
+GlassBR$(TARGET_EXTENSION): InputParameters.hpp InputFormat.hpp DerivedValues.hpp InputConstraints.hpp OutputFormat.hpp Calculations.hpp ReadTable.hpp Interpolation.hpp Control.cpp InputParameters.cpp InputFormat.cpp DerivedValues.cpp InputConstraints.cpp OutputFormat.cpp Calculations.cpp ReadTable.cpp Interpolation.cpp
+	"$(CXX)" Control.cpp InputParameters.cpp InputFormat.cpp DerivedValues.cpp InputConstraints.cpp OutputFormat.cpp Calculations.cpp ReadTable.cpp Interpolation.cpp --std=c++11 -o GlassBR$(TARGET_EXTENSION)
 
 run: build
 	./GlassBR$(TARGET_EXTENSION) $(RUNARGS)
