+++ code/stable/glassbr/src/csharp/Makefile
@@ -15,8 +15,8 @@ endif
 
 build: GlassBR$(TARGET_EXTENSION)
 
-GlassBR$(TARGET_EXTENSION): Control.cs InputParameters.cs DerivedValues.cs InputConstraints.cs InputFormat.cs OutputFormat.cs Calculations.cs ReadTable.cs Interpolation.cs
-	$(CSC) -out:GlassBR$(TARGET_EXTENSION) Control.cs InputParameters.cs DerivedValues.cs InputConstraints.cs InputFormat.cs OutputFormat.cs Calculations.cs ReadTable.cs Interpolation.cs
+GlassBR$(TARGET_EXTENSION): Control.cs InputParameters.cs InputFormat.cs DerivedValues.cs InputConstraints.cs OutputFormat.cs Calculations.cs ReadTable.cs Interpolation.cs
+	$(CSC) -out:GlassBR$(TARGET_EXTENSION) Control.cs InputParameters.cs InputFormat.cs DerivedValues.cs InputConstraints.cs OutputFormat.cs Calculations.cs ReadTable.cs Interpolation.cs
 
 run: build
 	./GlassBR$(TARGET_EXTENSION) $(RUNARGS)
