+++ code/stable/glassbr/src/java/Makefile
@@ -1,6 +1,6 @@
 build: GlassBR/Control.class
 
-GlassBR/Control.class: GlassBR/Control.java GlassBR/InputParameters.java GlassBR/DerivedValues.java GlassBR/InputConstraints.java GlassBR/InputFormat.java GlassBR/OutputFormat.java GlassBR/Calculations.java GlassBR/ReadTable.java GlassBR/Interpolation.java
+GlassBR/Control.class: GlassBR/Control.java GlassBR/InputParameters.java GlassBR/InputFormat.java GlassBR/DerivedValues.java GlassBR/InputConstraints.java GlassBR/OutputFormat.java GlassBR/Calculations.java GlassBR/ReadTable.java GlassBR/Interpolation.java
 	javac GlassBR/Control.java
 
 run: build
