+++ code/stable/nopcm/src/cpp/Control.cpp
@@ -13,7 +13,6 @@ using std::vector;
 using std::ifstream;
 using std::ofstream;
 
-#include "InputFormat.hpp"
 #include "InputParameters.hpp"
 #include "OutputFormat.hpp"
