+++ code/stable/nopcm/src/cpp/InputParameters.cpp
@@ -15,6 +15,54 @@ using std::vector;
 using std::ifstream;
 using std::ofstream;
 
+void get_input(string filename, double &A_C, double &C_W, double &h_C, double &T_init, double &t_final, double &L, double &T_C, double &t_step, double &rho_W, double &D, double &A_tol, double &R_tol, double &T_W, double &E_W) {
+    ifstream infile;
+    infile.open(filename, std::fstream::in);
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile >> A_C;
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile >> C_W;
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile >> h_C;
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile >> T_init;
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile >> t_final;
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile >> L;
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile >> T_C;
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile >> t_step;
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile >> rho_W;
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile >> D;
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile >> A_tol;
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile >> R_tol;
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile >> T_W;
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile >> E_W;
+    infile.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    infile.close();
+}
+
 void input_constraints(double A_C, double C_W, double h_C, double T_init, double t_final, double L, double T_C, double t_step, double rho_W, double D, double T_W, double E_W) {
     if (!((A_C <= 100000))) {
         std::cout << "Warning: constraint violated" << std::endl;
