--- code/stable/nopcm/src/csharp/InputFormat.cs
@@ -1,42 +1,0 @@
-using System;
-using System.IO;
-using System.Collections;
-using System.Collections.Generic;
-
-public class InputFormat {
-    
-    public static void get_input(string filename, out double A_C, out double C_W, out double h_C, out double T_init, out double t_final, out double L, out double T_C, out double t_step, out double rho_W, out double D, out double A_tol, out double R_tol, out double T_W, out double E_W) {
-        StreamReader infile;
-        infile = new StreamReader(filename);
-        infile.ReadLine();
-        A_C = Double.Parse(infile.ReadLine());
-        infile.ReadLine();
-        C_W = Double.Parse(infile.ReadLine());
-        infile.ReadLine();
-        h_C = Double.Parse(infile.ReadLine());
-        infile.ReadLine();
-        T_init = Double.Parse(infile.ReadLine());
-        infile.ReadLine();
-        t_final = Double.Parse(infile.ReadLine());
-        infile.ReadLine();
-        L = Double.Parse(infile.ReadLine());
-        infile.ReadLine();
-        T_C = Double.Parse(infile.ReadLine());
-        infile.ReadLine();
-        t_step = Double.Parse(infile.ReadLine());
-        infile.ReadLine();
-        rho_W = Double.Parse(infile.ReadLine());
-        infile.ReadLine();
-        D = Double.Parse(infile.ReadLine());
-        infile.ReadLine();
-        A_tol = Double.Parse(infile.ReadLine());
-        infile.ReadLine();
-        R_tol = Double.Parse(infile.ReadLine());
-        infile.ReadLine();
-        T_W = Double.Parse(infile.ReadLine());
-        infile.ReadLine();
-        E_W = Double.Parse(infile.ReadLine());
-        infile.Close();
-    }
-}
-
