+++ code/stable/nopcm/src/csharp/Makefile
@@ -15,8 +15,8 @@ endif
 
 build: SWHS$(TARGET_EXTENSION)
 
-SWHS$(TARGET_EXTENSION): Control.cs InputParameters.cs InputFormat.cs OutputFormat.cs
-	$(CSC) -out:SWHS$(TARGET_EXTENSION) Control.cs InputParameters.cs InputFormat.cs OutputFormat.cs
+SWHS$(TARGET_EXTENSION): Control.cs InputParameters.cs OutputFormat.cs
+	$(CSC) -out:SWHS$(TARGET_EXTENSION) Control.cs InputParameters.cs OutputFormat.cs
 
 run: build
 	./SWHS$(TARGET_EXTENSION) $(RUNARGS)
