+++ code/stable/nopcm/src/java/SWHS/InputParameters.java
@@ -10,6 +10,72 @@ import java.util.ArrayList;
 
 public class InputParameters {
     
+    public static Object[] get_input(String filename) throws Exception {
+        double A_C;
+        double C_W;
+        double h_C;
+        double T_init;
+        double t_final;
+        double L;
+        double T_C;
+        double t_step;
+        double rho_W;
+        double D;
+        double A_tol;
+        double R_tol;
+        double T_W;
+        double E_W;
+        
+        Scanner infile;
+        infile = new Scanner(new File(filename));
+        infile.nextLine();
+        A_C = Double.parseDouble(infile.nextLine());
+        infile.nextLine();
+        C_W = Double.parseDouble(infile.nextLine());
+        infile.nextLine();
+        h_C = Double.parseDouble(infile.nextLine());
+        infile.nextLine();
+        T_init = Double.parseDouble(infile.nextLine());
+        infile.nextLine();
+        t_final = Double.parseDouble(infile.nextLine());
+        infile.nextLine();
+        L = Double.parseDouble(infile.nextLine());
+        infile.nextLine();
+        T_C = Double.parseDouble(infile.nextLine());
+        infile.nextLine();
+        t_step = Double.parseDouble(infile.nextLine());
+        infile.nextLine();
+        rho_W = Double.parseDouble(infile.nextLine());
+        infile.nextLine();
+        D = Double.parseDouble(infile.nextLine());
+        infile.nextLine();
+        A_tol = Double.parseDouble(infile.nextLine());
+        infile.nextLine();
+        R_tol = Double.parseDouble(infile.nextLine());
+        infile.nextLine();
+        T_W = Double.parseDouble(infile.nextLine());
+        infile.nextLine();
+        E_W = Double.parseDouble(infile.nextLine());
+        infile.close();
+        
+        Object[] outputs = new Object[14];
+        outputs[0] = A_C;
+        outputs[1] = C_W;
+        outputs[2] = h_C;
+        outputs[3] = T_init;
+        outputs[4] = t_final;
+        outputs[5] = L;
+        outputs[6] = T_C;
+        outputs[7] = t_step;
+        outputs[8] = rho_W;
+        outputs[9] = D;
+        outputs[10] = A_tol;
+        outputs[11] = R_tol;
+        outputs[12] = T_W;
+        outputs[13] = E_W;
+        return outputs;
+    }
+    
     public static void input_constraints(double A_C, double C_W, double h_C, double T_init, double t_final, double L, double T_C, double t_step, double rho_W, double D, double T_W, double E_W) throws Exception {
         if (!((A_C <= 100000))) {
             System.out.println("Warning: constraint violated");
