+++ code/stable/projectile/src/cpp/Makefile
@@ -12,8 +12,8 @@ endif
 
 build: Projectile$(TARGET_EXTENSION)
 
-Projectile$(TARGET_EXTENSION): InputParameters.hpp InputConstraints.hpp InputFormat.hpp OutputFormat.hpp Calculations.hpp Control.cpp InputParameters.cpp InputConstraints.cpp InputFormat.cpp OutputFormat.cpp Calculations.cpp
-	"$(CXX)" Control.cpp InputParameters.cpp InputConstraints.cpp InputFormat.cpp OutputFormat.cpp Calculations.cpp --std=c++11 -o Projectile$(TARGET_EXTENSION)
+Projectile$(TARGET_EXTENSION): InputParameters.hpp InputFormat.hpp InputConstraints.hpp OutputFormat.hpp Calculations.hpp Control.cpp InputParameters.cpp InputFormat.cpp InputConstraints.cpp OutputFormat.cpp Calculations.cpp
+	"$(CXX)" Control.cpp InputParameters.cpp InputFormat.cpp InputConstraints.cpp OutputFormat.cpp Calculations.cpp --std=c++11 -o Projectile$(TARGET_EXTENSION)
 
 run: build
 	./Projectile$(TARGET_EXTENSION) $(RUNARGS)
