+++ code/stable/projectile/src/java/Makefile
@@ -1,6 +1,6 @@
 build: Projectile/Control.class
 
-Projectile/Control.class: Projectile/Control.java Projectile/InputParameters.java Projectile/InputConstraints.java Projectile/InputFormat.java Projectile/OutputFormat.java Projectile/Calculations.java
+Projectile/Control.class: Projectile/Control.java Projectile/InputParameters.java Projectile/InputFormat.java Projectile/InputConstraints.java Projectile/OutputFormat.java Projectile/Calculations.java
 	javac Projectile/Control.java
 
 run: build
